v2 - bugs fixed - new project structure
