+++ .gitignore
@@ -11,5 +11,5 @@ __pycache__/*
 env/
 venv/
 ENV/
-aluure-results/
-aluure-report/
+allure-results/
+allure-report/
