+++ locators/__init__.py
@@ -1,0 +1,2 @@
+from .add_customer_locators import AddCustomerLocators
+from .customers_locators import CustomersLocators
