+++ pages/__init__.py
@@ -1,0 +1,3 @@
+from .base_page import BasePage
+from .add_customer_page import AddCustomerPage
+from .customers_page import CustomersPage
