+++ run_task_ui.py
@@ -1,3 +1,4 @@
+import platform
 from os import system as os_system
 from os.path import join as os_path_join
 from webbrowser import open as open_report
@@ -6,10 +7,10 @@ from webbrowser import open as open_report
 def main():
     os_system('pytest tests/ -n 3 --alluredir=allure-results --clean-alluredir')
 
-    env_data = ['os_platform = Linux\n',
-                'os_release = 6.8.0-40-generic\n',
-                'os_version = Ubuntu 24.04 LTS\n',
-                'python_version = Python 3.10.14']
+    env_data = [f'os_platform = {platform.system()}\n',
+                f'os_release = {platform.release()}\n',
+                f'python_version = {platform.python_version()}']
+
     with open(os_path_join('allure-results', 'environment.properties'), 'w') as env_file:
         for line in env_data:
             env_file.write(line)
